% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_icon.R
\name{add_icon}
\alias{add_icon}
\title{add_icon}
\usage{
add_icon(id = NULL, x = 0, y = 0)
}
\arguments{
\item{id}{additional plot identificaition (URL, database ID, etc)}

\item{x}{x-coordinate of logo}

\item{y}{y-coordinate of logo}
}
\description{
add_icon
}
\author{
Mike Dietze
}
