% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_status_page.R
\name{create_status_page}
\alias{create_status_page}
\title{Create PEcAn Network status page.}
\usage{
create_status_page(config_file, file_prefix = "status", delta = 3600)
}
\arguments{
\item{config_file}{Path to `config.php`}

\item{file_prefix}{prefix used for all files saved}

\item{delta}{number of seconds before}
}
\description{
This will create a webpage (file_prefix.html) as well as an image
(file_prefix.png) that shows the PEcAn Network satus. This will
save the geo location of each site as well as some additional
information in a cache file (file_prefix.RData). If an update to
the BETY schema is detected it will be logged in a file
(file_prefix.log).
}
\examples{
\dontrun{
  create_status_page('/home/carya/pecan/web/config.php')
}
}
\author{
Michael Dietze

Rob Kooper
}
