% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_netcdf.R
\name{data.fetch}
\alias{data.fetch}
\title{data.fetch}
\usage{
data.fetch(var, nc, fun = mean)
}
\arguments{
\item{var}{the variable to extract from the hdf data}

\item{nc}{ncdf file path}

\item{fun}{the function to apply to the data at the same time, DEFAULT fun = mean}
}
\value{
aggregated data
}
\description{
data.fetch
}
