% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.r
\name{decile}
\alias{decile}
\title{deciling function}
\usage{
decile(x)
}
\arguments{
\item{x}{A numeric variable}
}
\value{
It will give back series of numbers, in which decile the observation goes to.
}
\description{
Takes the data and divide in to 10 parts.
}
\examples{
y<-round(runif(100,min=0,max=100))
y
x<-decile(y)
table(x)
x

}
\author{
Rajesh Jakhotia
}
