% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_data}
\alias{plot_data}
\alias{plot.data}
\title{Add data to plot}
\usage{
plot_data(trait.data, base.plot = NULL, ymax)
}
\arguments{
\item{trait.data}{Data to be plotted}

\item{base.plot}{a ggplot object (grob),
created if none provided}

\item{ymax}{maximum height of y}
}
\value{
updated plot object
}
\description{
Add data to an existing plot or create a new one
}
\details{
Used to add raw data or summary statistics to the plot of a distribution.
The height of Y is arbitrary, and can be set to optimize visualization.
If SE estimates are available, the SE will be plotted
}
\examples{
\dontrun{plot_data(data.frame(Y = c(1, 2), se = c(1,2)), base.plot = NULL, ymax = 10)}
}
\author{
David LeBauer
}
