% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_netcdf.R
\name{plot_netcdf}
\alias{plot_netcdf}
\alias{plot.netcdf}
\title{Load the tower dataset and create a plot.}
\usage{
plot_netcdf(
  datafile,
  yvar,
  xvar = "time",
  width = 800,
  height = 600,
  filename = NULL,
  year = NULL
)
}
\arguments{
\item{datafile}{the specific datafile to use.}

\item{yvar}{the variable to plot along the y-axis.}

\item{xvar}{the variable to plot along the x-axis, by default time is
used.}

\item{width}{the width of the image generated, default is 800 pixels.}

\item{height}{the height of the image generated, default is 600 pixels.}

\item{filename}{is the name of the file name that is geneated, this
can be null to use existing device, otherwise it will try and
create an image based on filename, or display if x11.}

\item{year}{the year this data is for (only used in the title).}
}
\description{
Loads the tower data from an HDF5 file generated by
ED and will plot the values against one another. The default is for
the given variable to be plotted against time.
}
\author{
Rob Kooper
}
