% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.r
\name{Squire}
\alias{Squire}
\title{Squire Function}
\usage{
Squire(x)
}
\arguments{
\item{x}{numeric variable}
}
\value{
x^2
}
\description{
Takes numeric in puts x
}
\examples{
Squire(2)
}
\author{
Jishnu
}
