% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciEnvelope.R
\name{ciEnvelope}
\alias{ciEnvelope}
\title{plots a confidence interval around an x-y plot (e.g. a timeseries)}
\usage{
ciEnvelope(x, ylo, yhi, ...)
}
\arguments{
\item{x}{Vector defining CI center}

\item{ylo}{Vector defining bottom of CI envelope}

\item{yhi}{Vector defining top of CI envelope}

\item{...}{further arguments passed on to `graphics::polygon`}
}
\description{
plots a confidence interval around an x-y plot (e.g. a timeseries)
}
\author{
Michael Dietze, David LeBauer
}
