% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map.output.R
\name{map.output}
\alias{map.output}
\title{Map Output}
\usage{
map.output(table, variable)
}
\arguments{
\item{table}{data.table or data.frame with columns lat, lon, followed by variable names}

\item{variable}{name of variable to be mapped}
}
\value{
plot
}
\description{
Map Output
}
\author{
David LeBauer
}

